+++ src/system/hal/common/hal.h
@@ -43,28 +43,28 @@ typedef volatile uint32_t* mmio_t;
  * @param mask		The value to write
  * @param size		The size of the register
  */
-uint32_t hal_bitmask_write(uint32_t port, uint8_t offset, uint32_t mask, uint8_t size);
+uint32_t hal_bitmask_write(uint32_t port, uint32_t offset, uint32_t mask, uint32_t size);
 
 /**
  * Enable a bit(mask) of given port and offset
  */
-void hal_bitmask_set(uint32_t port, uint8_t offset, uint32_t mask);
+void hal_bitmask_set(uint32_t port, uint32_t offset, uint32_t mask);
 
 /**
  * Clear a bit(mask) of given port and offset
  */
-void hal_bitmask_clear(uint32_t port, uint8_t offset, uint32_t mask);
+void hal_bitmask_clear(uint32_t port, uint32_t offset, uint32_t mask);
 
 /**
  * Get value of given address
  */
-uint32_t hal_get_address_value(uint32_t port, uint8_t offset);
+uint32_t hal_get_address_value(uint32_t port, uint32_t offset);
 
 /**
  * Returns the register address of given port and offset
  *
  * @internal
  */
-inline mmio_t hal_get_register(uint32_t port, uint8_t offset);
+inline mmio_t hal_get_register(uint32_t port, uint32_t offset);
 
 #endif /* SRC_SYSTEM_HAL_HAL_H_ */
